pem_rfc7468

Module grammar

Helper functions and rules for enforcing the ABNF grammar for RFC 7468-flavored PEM as described in Section 3.

The grammar described below is intended to follow the “ABNF (Strict)” subset of the grammar as described in Section 3 Figure 3.

Constants

CHAR_COLON 🔒

Colon ‘:’

CHAR_CR 🔒

Carriage return

CHAR_HT 🔒

Horizontal tab

CHAR_LF 🔒

Line feed

CHAR_NUL 🔒

NUL char

CHAR_SP 🔒

Space

Functions

is_allowed_in_label 🔒

Does the provided byte match a character allowed in a label?

is_labelchar 🔒

Any printable character except hyphen-minus, as defined in the ‘labelchar’ production in the RFC 7468 ABNF grammar

is_wsp 🔒

Does the provided byte match the “WSP” ABNF production from Section 3?

split_label 🔒

Split a slice beginning with a type label as located in an encapsulation boundary. Returns the label as a &str, and slice beginning with the encapsulated text with leading ----- and newline removed.

strip_leading_eol 🔒

Strip a newline (eol) from the beginning of the provided byte slice.

strip_preamble 🔒

Strip the “preamble”, i.e. data that appears before the PEM pre-encapsulation boundary.

strip_trailing_eol 🔒

Strip a newline (eol) from the end of the provided byte slice.

validate_label 🔒

Validate that the given bytes are allowed as a PEM type label, i.e. the label encoded in the BEGIN and END encapsulation boundaries.